refactor: simplify bank import dedupe helpers
